@@ -1,47 +1,48 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFramework>net461</TargetFramework>
-		<LangVersion>9</LangVersion>
+		<TargetFramework>net7</TargetFramework>
+		<LangVersion>11</LangVersion>
 	</PropertyGroup>
-	
-  <ItemGroup>
-		<Reference Include="Vintagestory">
-				<HintPath>$(VINTAGE_STORY)\Vintagestory.exe</HintPath>
-				<Private>false</Private>
-		</Reference>
-		<Reference Include="VintagestoryAPI">
-				<HintPath>$(VINTAGE_STORY)\VintagestoryAPI.dll</HintPath>
-				<Private>false</Private>
-		</Reference>
-		<Reference Include="VintagestoryLib">
-				<HintPath>$(VINTAGE_STORY)\VintagestoryLib.dll</HintPath>
-				<Private>false</Private>
-		</Reference>
-		<Reference Include="VSCreativeMod">
-				<HintPath>$(VINTAGE_STORY)\Mods\VSCreativeMod.dll</HintPath>
-				<Private>false</Private>
-		</Reference>
-		<Reference Include="VSSurvivalMod">
-				<HintPath>$(VINTAGE_STORY)\Mods\VSSurvivalMod.dll</HintPath>
-				<Private>false</Private>
-		</Reference>
-		<Reference Include="VSEssentials">
-				<HintPath>$(VINTAGE_STORY)\Mods\VSEssentials.dll</HintPath>
-				<Private>false</Private>
-		</Reference>
-    <Reference Include="Newtonsoft.Json, Version=8.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>$(VINTAGE_STORY)\Lib\Newtonsoft.Json.dll</HintPath>
-      <Private>false</Private>
-    </Reference>
-  </ItemGroup>
+
+	<ItemGroup>
+        <Reference Include="Vintagestory">
+            <HintPath>$(VINTAGE_STORY)\Vintagestory.exe</HintPath>
+            <Private>false</Private>
+        </Reference>
+        <Reference Include="VintagestoryAPI">
+            <HintPath>$(VINTAGE_STORY)\VintagestoryAPI.dll</HintPath>
+            <Private>false</Private>
+        </Reference>
+        <Reference Include="VintagestoryLib">
+            <HintPath>$(VINTAGE_STORY)\VintagestoryLib.dll</HintPath>
+            <Private>false</Private>
+        </Reference>
+        <Reference Include="VSSurvivalMod">
+            <HintPath>$(VINTAGE_STORY)\Mods\VSSurvivalMod.dll</HintPath>
+            <Private>false</Private>
+        </Reference>
+        <Reference Include="VSEssentials">
+            <HintPath>$(VINTAGE_STORY)\Mods\VSEssentials.dll</HintPath>
+            <Private>false</Private>
+        </Reference>
+        <!-- Vintage Story Included Libraries -->
+        <Reference Include="0Harmony">
+            <HintPath>$(VINTAGE_STORY)\Lib\0Harmony.dll</HintPath>
+            <Private>false</Private>
+        </Reference>
+        <Reference Include="cairo-sharp">
+            <HintPath>$(VINTAGE_STORY)\Lib\cairo-sharp.dll</HintPath>
+            <Private>false</Private>
+        </Reference>
+        <Reference Include="Newtonsoft.Json">
+            <HintPath>$(VINTAGE_STORY)\Lib\Newtonsoft.Json.dll</HintPath>
+            <Private>false</Private>
+        </Reference>
+	</ItemGroup>
 
 	<ItemGroup Condition="'$(Configuration)' == 'ReleaseToMods'">
+		<None Include="*ICENSE" CopyToOutputDirectory="PreserveNewest" />
 		<None Include="resources/**">
-			<Link>%(RecursiveDir)%(Filename)%(Extension)</Link>
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</None>
-		<None Include="LICENSE">
 			<Link>%(RecursiveDir)%(Filename)%(Extension)</Link>
 			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
 		</None>
